+++ src/test/delete-db.test.ts
@@ -1,12 +1,4 @@
-import {
-  describe,
-  it,
-  beforeAll,
-  afterAll,
-  beforeEach,
-  afterEach,
-  expect,
-} from './test-utils.ts';
+import { describe, it, beforeAll, afterAll, expect } from './test-utils.ts';
 import { rm } from 'fs/promises';
 import { existsSync } from 'fs';